requirements.txt
streamlit
replicate
python-dotenv
pandas
plotly
requests